pyproject.toml
[tool.poetry]
name = "eabotbackend"
version = "0.1.0"
description = "Django backend for the Reddit EA Bot project"
authors = ["Andrés Arango Pérez <andresap@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.7"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
